requirements.txt
autopep8==1.5.4
click==7.1.2
Jinja2==2.11.2
markdown2==2.3.9
MarkupSafe==1.1.1
pycodestyle==2.6.0
PyYAML==5.3.1
toml==0.10.1
